Correct display of time in hours and minutes.
This change makes the preparation and cooking time being correctly displayed online.
The times which are stored in minutes are now correctly reported as hours+minutes.

Also not a small change in the definition of `TimesInMinsAdapted`, to ensure that
the time are reported as integers instead od doubles on the display.