Introduce Recipe Preparation
This small change introduces an end-point with an idempotent operation that
is going to convert about 30 recipes from New to Ready.

The purpose is to prepare the database for another commit which is then going
to select them for curation and present them to users instead of recipes
in New status.

The commit after that will upgrade the preparation to include more
recipes (possibly recipes upgraded from the original partial parsing).